fix: set correct visibility to plugin configuration
